OpenStack cluster (#13)
* add openstack cluster


* cluster version upgrade


* assign sshkey to clushter

+++ kubermatic/resource_cluster.go
@@ -4,8 +4,11 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/hashicorp/go-version"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/customdiff"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
 	"github.com/kubermatic/go-kubermatic/client/project"
 	"github.com/kubermatic/go-kubermatic/models"
 )
@@ -25,17 +28,31 @@ func resourceCluster() *schema.Resource {
 			"project_id": {
 				Type:        schema.TypeString,
 				Required:    true,
+				ForceNew:    true,
 				Description: "Reference project identifier",
 			},
 			"dc": {
 				Type:        schema.TypeString,
 				Required:    true,
+				ForceNew:    true,
 				Description: "Data center name",
 			},
 			"name": {
 				Type:        schema.TypeString,
 				Required:    true,
 				Description: "Cluster name",
+			},
+			"labels": {
+				Type:     schema.TypeMap,
+				Optional: true,
+			},
+			"sshkeys": {
+				Type:     schema.TypeSet,
+				Optional: true,
+				Elem: &schema.Schema{
+					Type:         schema.TypeString,
+					ValidateFunc: validation.NoZeroValues,
+				},
 			},
 			"spec": {
 				Type:        schema.TypeList,
@@ -54,6 +71,7 @@ func resourceCluster() *schema.Resource {
 			"type": {
 				Type:        schema.TypeString,
 				Optional:    true,
+				ForceNew:    true,
 				Default:     "kubernetes",
 				Description: "Cluster type Kubernetes or OpenShift",
 			},
@@ -68,6 +86,27 @@ func resourceCluster() *schema.Resource {
 				Description: "Deletion timestamp",
 			},
 		},
+
+		CustomizeDiff: customdiff.All(
+			customdiff.ForceNewIfChange("spec.0.version", func(old, new, meta interface{}) bool {
+				// "version" can only be upgraded to newer versions, so we must create a new resource
+				// if it is decreased.
+				newVer, err := version.NewVersion(new.(string))
+				if err != nil {
+					return false
+				}
+
+				oldVer, err := version.NewVersion(old.(string))
+				if err != nil {
+					return false
+				}
+
+				if newVer.LessThan(oldVer) {
+					return true
+				}
+				return false
+			}),
+		),
 	}
 }
 
@@ -84,6 +123,7 @@ func resourceClusterCreate(d *schema.ResourceData, m interface{}) error {
 			Name:       d.Get("name").(string),
 			Spec:       expandClusterSpec(d.Get("spec").([]interface{})),
 			Type:       d.Get("type").(string),
+			Labels:     getLabels(d),
 			Credential: d.Get("credential").(string),
 		},
 	})
@@ -93,37 +133,24 @@ func resourceClusterCreate(d *schema.ResourceData, m interface{}) error {
 		return fmt.Errorf("unable to create cluster for project '%s': %s", pID, err)
 	}
 	d.SetId(r.Payload.ID)
-	cID := r.Payload.ID
 
-	err = resource.Retry(d.Timeout(schema.TimeoutCreate), func() *resource.RetryError {
-		hp := project.NewGetClusterHealthParams()
-		hp.SetClusterID(cID)
-		hp.SetProjectID(pID)
-		hp.SetDC(dc)
-
-		r, err := k.client.Project.GetClusterHealth(hp, k.auth)
-		if err != nil {
-			return resource.NonRetryableError(fmt.Errorf("unable to get cluster '%s' health: %v", cID, err))
-		}
-
-		if r.Payload.Apiserver == healthStatusUp &&
-			r.Payload.CloudProviderInfrastructure == healthStatusUp &&
-			r.Payload.Controller == healthStatusUp &&
-			r.Payload.Etcd == healthStatusUp &&
-			r.Payload.MachineController == healthStatusUp &&
-			r.Payload.Scheduler == healthStatusUp &&
-			r.Payload.UserClusterControllerManager == healthStatusUp {
-			return nil
-		}
-
-		k.log.Debugf("waiting for cluster '%s' to be ready, %+v", cID, r.Payload)
-		return resource.RetryableError(fmt.Errorf("waiting for cluster '%s' to be ready", cID))
-	})
-	if err != nil {
-		return fmt.Errorf("cluster '%s' is not ready: %v", cID, err)
+	if err := waitClusterReady(k, d); err != nil {
+		return fmt.Errorf("cluster '%s' is not ready: %v", r.Payload.ID, err)
 	}
 
 	return resourceClusterRead(d, m)
+}
+
+func getLabels(d *schema.ResourceData) map[string]string {
+	var labels map[string]string
+	if v := d.Get("labels"); v != nil {
+		labels = make(map[string]string)
+		m := d.Get("labels").(map[string]interface{})
+		for k, v := range m {
+			labels[k] = v.(string)
+		}
+	}
+	return labels
 }
 
 func resourceClusterRead(d *schema.ResourceData, m interface{}) error {
@@ -149,6 +176,14 @@ func resourceClusterRead(d *schema.ResourceData, m interface{}) error {
 		return fmt.Errorf("unable to get cluster '%s': %v", d.Id(), err)
 	}
 
+	labels, err := excludeProjectLabels(k, d.Get("project_id").(string), r.Payload.Labels)
+	if err != nil {
+		return err
+	}
+	if err := d.Set("labels", labels); err != nil {
+		return err
+	}
+
 	d.Set("name", r.Payload.Name)
 
 	// TODO: check why API returns an empty credential field even if it is set
@@ -159,9 +194,9 @@ func resourceClusterRead(d *schema.ResourceData, m interface{}) error {
 
 	d.Set("type", r.Payload.Type)
 
-	// TODO: Rewriting sensitive fields produces misleading state diff, which
-	// can be fixed by skipping and not rewriting sensitive fields.
-	if err = d.Set("spec", flattenClusterSpec(r.Payload.Spec)); err != nil {
+	values := readClusterPreserveValues(d)
+	specFlattenned := flattenClusterSpec(values, r.Payload.Spec)
+	if err = d.Set("spec", specFlattenned); err != nil {
 		return err
 	}
 
@@ -169,13 +204,211 @@ func resourceClusterRead(d *schema.ResourceData, m interface{}) error {
 
 	d.Set("deletion_timestamp", r.Payload.DeletionTimestamp.String())
 
+	keys, err := getClusterAssignedSSHKeys(d, k)
+	if err != nil {
+		return err
+	}
+	if err := d.Set("sshkeys", keys); err != nil {
+		return err
+	}
+
 	return nil
 }
 
+// excludeProjectLabels excludes labels defined in project.
+// Project labels propogated to clusters. For better predictability of
+// cluster's labels changes, project's labels are excluded from cluster state.
+func excludeProjectLabels(k *kubermaticProviderMeta, projectID string, allLabels map[string]string) (map[string]string, error) {
+	p := project.NewGetProjectParams()
+	p.SetProjectID(projectID)
+
+	r, err := k.client.Project.GetProject(p, k.auth)
+	if err != nil {
+		return nil, err
+	}
+
+	for k := range r.Payload.Labels {
+		delete(allLabels, k)
+	}
+
+	return allLabels, nil
+}
+
+func getClusterAssignedSSHKeys(d *schema.ResourceData, k *kubermaticProviderMeta) ([]string, error) {
+	p := project.NewListSSHKeysAssignedToClusterParams()
+	p.SetProjectID(d.Get("project_id").(string))
+	p.SetDC(d.Get("dc").(string))
+	p.SetClusterID(d.Id())
+	ret, err := k.client.Project.ListSSHKeysAssignedToCluster(p, k.auth)
+	if err != nil {
+		return nil, err
+	}
+
+	var ids []string
+	for _, v := range ret.Payload {
+		ids = append(ids, v.ID)
+	}
+	return ids, nil
+}
+
+// clusterPreserveValues helps avoid misleading diffs during read phase.
+// API result does not have some important fields valeus, like sensitive
+// access key or password fields. When API result is flattened and written to
+// terraform state it creating state diff that might force undesired updates and
+// even force replacement of a cluster. Solution is to set values for preserved
+// values in flattened object before comitting it to state.
+type clusterPreserveValues struct {
+	openstackUsername interface{}
+	openstackPassword interface{}
+	openstackTenant   interface{}
+}
+
+func readClusterPreserveValues(d *schema.ResourceData) clusterPreserveValues {
+	return clusterPreserveValues{
+		openstackUsername: d.Get("spec.0.cloud.0.openstack.0.username"),
+		openstackPassword: d.Get("spec.0.cloud.0.openstack.0.password"),
+		openstackTenant:   d.Get("spec.0.cloud.0.openstack.0.tenant"),
+	}
+}
+
 func resourceClusterUpdate(d *schema.ResourceData, m interface{}) error {
-	// TODO: implement after kubermatic client fix
+	d.Partial(true)
+	defer d.Partial(false)
+
+	k := m.(*kubermaticProviderMeta)
+
+	if d.HasChanges("name", "labels", "spec") {
+		patchClusterFields(d, k)
+	}
+	if d.HasChange("sshkeys") {
+		updateClusterSSHKeys(d, k)
+	}
+
+	if err := waitClusterReady(k, d); err != nil {
+		return fmt.Errorf("cluster '%s' is not ready: %v", d.Id(), err)
+	}
 
 	return resourceClusterRead(d, m)
+}
+
+func patchClusterFields(d *schema.ResourceData, k *kubermaticProviderMeta) error {
+	p := project.NewPatchClusterParams()
+	p.SetProjectID(d.Get("project_id").(string))
+	p.SetDC(d.Get("dc").(string))
+	p.SetClusterID(d.Id())
+	name := d.Get("name").(string)
+	version := d.Get("spec.0.version").(string)
+	auditLogging := d.Get("spec.0.audit_logging").(bool)
+	labels := d.Get("labels")
+	p.SetPatch(newClusterPatch(name, version, auditLogging, labels))
+
+	err := resource.Retry(d.Timeout(schema.TimeoutUpdate), func() *resource.RetryError {
+		_, err := k.client.Project.PatchCluster(p, k.auth)
+		if err != nil {
+			if e, ok := err.(*project.PatchClusterDefault); ok && e.Code() == http.StatusConflict {
+				return resource.RetryableError(fmt.Errorf("cluster patch conflict: %w", err))
+			}
+			return resource.NonRetryableError(fmt.Errorf("patch cluster '%s': %v", d.Id(), err))
+		}
+		return nil
+	})
+	if err != nil {
+		return err
+	}
+
+	d.SetPartial("name")
+	d.SetPartial("labels")
+	d.SetPartial("spec")
+
+	return nil
+}
+
+func updateClusterSSHKeys(d *schema.ResourceData, k *kubermaticProviderMeta) error {
+	var unassign, assign []string
+	old, new := d.GetChange("sshkeys")
+
+	for _, id := range old.(*schema.Set).List() {
+		if !new.(*schema.Set).Contains(id) {
+			unassign = append(unassign, id.(string))
+		}
+	}
+
+	for _, id := range new.(*schema.Set).List() {
+		if !old.(*schema.Set).Contains(id) {
+			assign = append(assign, id.(string))
+		}
+	}
+
+	projectID := d.Get("project_id").(string)
+	dc := d.Get("dc").(string)
+
+	for _, id := range unassign {
+		p := project.NewDetachSSHKeyFromClusterParams()
+		p.SetProjectID(projectID)
+		p.SetDC(dc)
+		p.SetClusterID(d.Id())
+		p.SetKeyID(id)
+		_, err := k.client.Project.DetachSSHKeyFromCluster(p, k.auth)
+		if err != nil {
+			return err
+		}
+	}
+
+	for _, id := range assign {
+		p := project.NewAssignSSHKeyToClusterParams()
+		p.SetProjectID(projectID)
+		p.SetDC(dc)
+		p.SetClusterID(d.Id())
+		p.SetKeyID(id)
+		_, err := k.client.Project.AssignSSHKeyToCluster(p, k.auth)
+		if err != nil {
+			return err
+		}
+	}
+
+	d.SetPartial("sshkeys")
+	return nil
+}
+
+func waitClusterReady(k *kubermaticProviderMeta, d *schema.ResourceData) error {
+	return resource.Retry(d.Timeout(schema.TimeoutCreate), func() *resource.RetryError {
+		hp := project.NewGetClusterHealthParams()
+		hp.SetClusterID(d.Id())
+		hp.SetProjectID(d.Get("project_id").(string))
+		hp.SetDC(d.Get("dc").(string))
+
+		r, err := k.client.Project.GetClusterHealth(hp, k.auth)
+		if err != nil {
+			return resource.NonRetryableError(fmt.Errorf("unable to get cluster '%s' health: %v", d.Id(), err))
+		}
+
+		if r.Payload.Apiserver == healthStatusUp &&
+			r.Payload.CloudProviderInfrastructure == healthStatusUp &&
+			r.Payload.Controller == healthStatusUp &&
+			r.Payload.Etcd == healthStatusUp &&
+			r.Payload.MachineController == healthStatusUp &&
+			r.Payload.Scheduler == healthStatusUp &&
+			r.Payload.UserClusterControllerManager == healthStatusUp {
+			return nil
+		}
+
+		k.log.Debugf("waiting for cluster '%s' to be ready, %+v", d.Id(), r.Payload)
+		return resource.RetryableError(fmt.Errorf("waiting for cluster '%s' to be ready", d.Id()))
+	})
+}
+
+func newClusterPatch(name, version string, auditLogging bool, labels interface{}) interface{} {
+	// TODO(furkhat): change to dedicated struct when API has it.
+	return map[string]interface{}{
+		"name":   name,
+		"labels": labels,
+		"spec": map[string]interface{}{
+			"auditLogging": map[string]bool{
+				"enabled": auditLogging,
+			},
+			"version": version,
+		},
+	}
 }
 
 func resourceClusterDelete(d *schema.ResourceData, m interface{}) error {
@@ -189,12 +422,26 @@ func resourceClusterDelete(d *schema.ResourceData, m interface{}) error {
 	p.SetProjectID(pID)
 	p.SetClusterID(cID)
 
-	_, err := k.client.Project.DeleteCluster(p, k.auth)
-	if err != nil {
-		return fmt.Errorf("unable to delete cluster '%s': %v", cID, err)
-	}
-
+	deleteSent := false
 	return resource.Retry(d.Timeout(schema.TimeoutDelete), func() *resource.RetryError {
+		if !deleteSent {
+			_, err := k.client.Project.DeleteCluster(p, k.auth)
+			if err != nil {
+				if e, ok := err.(*project.DeleteClusterDefault); ok {
+					if e.Code() == http.StatusConflict {
+						return resource.RetryableError(err)
+					}
+					if e.Code() == http.StatusNotFound {
+						return nil
+					}
+				}
+				if _, ok := err.(*project.DeleteClusterForbidden); ok {
+					return resource.RetryableError(err)
+				}
+				return resource.NonRetryableError(fmt.Errorf("unable to delete cluster '%s': %v", cID, err))
+			}
+			deleteSent = true
+		}
 		p := project.NewGetClusterParams()
 
 		p.SetClusterID(cID)
@@ -206,7 +453,10 @@ func resourceClusterDelete(d *schema.ResourceData, m interface{}) error {
 			if e, ok := err.(*project.GetClusterDefault); ok && e.Code() == http.StatusNotFound {
 				k.log.Debugf("cluster '%s' has been destroyed, returned http code: %d", cID, e.Code())
 				d.SetId("")
-				return nil
+				return resource.RetryableError(err)
+			}
+			if _, ok := err.(*project.GetClusterForbidden); ok {
+				return resource.RetryableError(err)
 			}
 			return resource.NonRetryableError(fmt.Errorf("unable to get cluster '%s': %v", cID, err))
 		}
